+++ rulewerk-core/src/test/java/org/semanticweb/rulewerk/core/reasoner/implementation/RdfFileDataSourceTest.java
@@ -1,4 +1,4 @@
-package org.semanticweb.rulewerk.reasoner.vlog;
+package org.semanticweb.rulewerk.core.reasoner.implementation;
 
 /*-
  * #%L
@@ -20,39 +20,32 @@ package org.semanticweb.rulewerk.reasoner.vlog;
  * #L%
  */
 
-import static org.junit.Assert.assertEquals;
-
 import java.io.File;
 import java.io.IOException;
 
 import org.junit.Test;
 
-import org.semanticweb.rulewerk.core.reasoner.implementation.FileDataSource;
-import org.semanticweb.rulewerk.core.reasoner.implementation.CsvFileDataSource;
+public class RdfFileDataSourceTest {
 
-public class CsvFileDataSourceTest {
-
-	private final String ntFile = FileDataSourceTestUtils.INPUT_FOLDER + "file.nt";
-	private final String csvFile = FileDataSourceTestUtils.INPUT_FOLDER + "file.csv";
-	private final String gzFile = csvFile + ".gz";
+	private final String unzippedRdfFile = FileDataSourceTestUtils.INPUT_FOLDER + "file.nt";
+	private final String zippedRdfFile = FileDataSourceTestUtils.INPUT_FOLDER + "file.nt.gz";
 
 	@Test(expected = NullPointerException.class)
 	public void testConstructorNullFile() throws IOException {
-		new CsvFileDataSource(null);
+		new RdfFileDataSource(null);
 	}
 
 	@Test(expected = IllegalArgumentException.class)
 	public void testConstructorFalseExtension() throws IOException {
-		new CsvFileDataSource(ntFile);
+		new RdfFileDataSource(FileDataSourceTestUtils.INPUT_FOLDER + "file.csv");
 	}
 
 	@Test
 	public void testConstructor() throws IOException {
-		final String dirCanonicalPath = new File(FileDataSourceTestUtils.INPUT_FOLDER).getCanonicalPath();
-		final CsvFileDataSource unzippedCsvFileDataSource = new CsvFileDataSource(csvFile);
-		final CsvFileDataSource zippedCsvFileDataSource = new CsvFileDataSource(gzFile);
+		final RdfFileDataSource unzippedRdfFileDataSource = new RdfFileDataSource(unzippedRdfFile);
+		final RdfFileDataSource zippedRdfFileDataSource = new RdfFileDataSource(zippedRdfFile);
 
-		FileDataSourceTestUtils.testConstructor(unzippedCsvFileDataSource, new File(csvFile).getName());
-		FileDataSourceTestUtils.testConstructor(zippedCsvFileDataSource, new File(gzFile).getName());
+		FileDataSourceTestUtils.testConstructor(unzippedRdfFileDataSource, new File(unzippedRdfFile).getName());
+		FileDataSourceTestUtils.testConstructor(zippedRdfFileDataSource, new File(zippedRdfFile).getName());
 	}
 }
