+++ rulewerk-core/src/test/java/org/semanticweb/rulewerk/core/reasoner/implementation/SparqlQueryResultDataSourceTest.java
@@ -1,4 +1,4 @@
-package org.semanticweb.rulewerk.reasoner.vlog;
+package org.semanticweb.rulewerk.core.reasoner.implementation;
 
 /*-
  * #%L
@@ -9,9 +9,9 @@ package org.semanticweb.rulewerk.reasoner.vlog;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -19,8 +19,6 @@ package org.semanticweb.rulewerk.reasoner.vlog;
  * limitations under the License.
  * #L%
  */
-
-import static org.junit.Assert.assertEquals;
 
 import java.io.IOException;
 import java.net.MalformedURLException;
@@ -32,8 +30,6 @@ import org.apache.commons.lang3.StringUtils;
 import org.junit.Test;
 import org.semanticweb.rulewerk.core.model.api.Variable;
 import org.semanticweb.rulewerk.core.model.implementation.Expressions;
-import org.semanticweb.rulewerk.core.reasoner.implementation.FileDataSource;
-import org.semanticweb.rulewerk.core.reasoner.implementation.SparqlQueryResultDataSource;
 
 public class SparqlQueryResultDataSourceTest {
